Authentication Finale (#12)
* Added the Promoter and Mana
ger users. Added the sign up routes for them. Added tests. Fixed a bug with UUID, changed to a normal primary key. Replaced the use of a firebase.json to a .env file. Changed the way the JSON is returned to only return desired fields. Added fields to the App User.

* Changed the .flake8 file so it excludes Django migrations code and other files. Increased the character limit per line to 120 so it matches with the line in PyCharm. Reformated my code with PyCharm and fixed some stuff so it passes the tests.

* Fixed the erros in the documentation

* Changed from HttpResponse to JsonResponse in the POST responses. Changed from square brackets to the get method to get an header field because the get defaults to None when the field is not provided. Changed the AppUser model to allow for some nullable fields.

* Removed the unused import.

* Add secrets to CI

Co-authored-by: Luís Alves <luismig.alves@gmail.com>